refactor : move to web repository
CLI 환경 추가에 따른 web 및 cli 분리
